dev-requirements.txt
django-nose
django-debug-toolbar
